Use a trait to protect nodes.
