+++ lib/Dav/System/Configuration/Node.php
@@ -21,6 +21,7 @@
  */
 namespace Sabre\Katana\Dav\System\Configuration;
 
+use Sabre\Katana\Dav\System;
 use Sabre\DAV as SabreDav;
 use Sabre\DAVACL as SabreDavAcl;
 
@@ -33,99 +34,12 @@ use Sabre\DAVACL as SabreDavAcl;
  */
 class Node extends SabreDav\Node implements SabreDavAcl\IACL {
 
+    use System\ProtectedNode;
+
     /**
      * Define the node name.
      *
      * @const string
      */
     const NAME = 'configurations';
-
-    /**
-     * Get the node's name.
-     *
-     * @return string
-     */
-    function getName() {
-
-        return self::NAME;
-    }
-
-    /**
-     * Get the owner principal
-     *
-     * This must be a url to a principal, or null if there's no owner.
-     *
-     * @return string|null
-     */
-    function getOwner() {
-
-        return 'principals/admin';
-    }
-
-    /**
-     * Get a group principal
-     *
-     * This must be a URL to a principal, or null if there's no owner.
-     *
-     * @return string|null
-     */
-    function getGroup() {
-
-        return null;
-    }
-
-    /**
-     * Get a list of ACE's for this node.
-     *
-     * Each ACE has the following properties:
-     *   * 'privilege', a string such as {DAV:}read or {DAV:}write. These are
-     *     currently the only supported privileges
-     *   * 'principal', a url to the principal who owns the node
-     *   * 'protected' (optional), indicating that this ACE is not allowed to
-     *      be updated.
-     *
-     * @return array
-     */
-    function getACL() {
-
-        return [
-            [
-                'principal' => $this->getOwner(),
-                'privilege' => '{DAV:}read',
-                'protected' => true
-            ]
-        ];
-    }
-
-    /**
-     * Updates the ACL.
-     *
-     * This method will receive a list of new ACE's as an array argument.
-     *
-     * @param array $acl
-     * @return void
-     */
-    function setACL(array $acl) {
-
-        throw new SabreDav\Exception\Forbidden(
-            'Updating ACLs it not allowed on this node.'
-        );
-    }
-
-    /**
-     * Get the list of supported privileges for this node.
-     *
-     * The returned data structure is a list of nested privileges.
-     * See Sabre\DAVACL\Plugin::getDefaultSupportedPrivilegeSet for a simple
-     * standard structure.
-     *
-     * If null is returned from this method, the default privilege set is used,
-     * which is fine for most common usecases.
-     *
-     * @return array|null
-     */
-    function getSupportedPrivilegeSet() {
-
-        return null;
-    }
 }
